+++ app/src/main/java/com/codesoom/assignment/domain/TaskRepository.java
@@ -1,41 +1,13 @@
 package com.codesoom.assignment.domain;
 
-import com.codesoom.assignment.TaskNotFoundException;
-import org.springframework.stereotype.Repository;
-
-import java.util.ArrayList;
 import java.util.List;
 
-@Repository
-public class TaskRepository {
+public abstract class TaskRepository {
+    public abstract List<Task> findAll();
 
-    private List<Task> tasks = new ArrayList<>();
-    private Long newId = 0L;
+    public abstract Task find(Long id);
 
-    public List<Task> findAll() {
-        return tasks;
-    }
+    public abstract Task save(Task task);
 
-    public Task find(Long id) {
-        return tasks.stream()
-                .filter(task -> task.getId().equals(id))
-                .findFirst()
-                .orElseThrow(() -> new TaskNotFoundException(id));
-    }
-
-    public Task save(Task task) {
-        task.setId(generateId());
-        tasks.add(task);
-        return task;
-    }
-
-    public Task remove(Task task) {
-        tasks.remove(task);
-        return task;
-    }
-
-    private Long generateId() {
-        newId += 1;
-        return newId;
-    }
+    public abstract Task remove(Task task);
 }
